+++ test/msh.py
@@ -31,14 +31,14 @@ class MSH:
         MSH.HOME.add_file(".conf", content=MSH.CONFIG).make()
 
     @staticmethod
-    def run_suc(command: str):
+    def run_suc(command: str) -> Log:
         suc, output = MSH.run(command)
         if not assert_true(suc, raise_assertion_error=False):
             raise AssertionError(f"Command '{command}' failed: \n{output}")
         return output
 
     @staticmethod
-    def run_fail(command: str):
+    def run_fail(command: str) -> Log:
         suc, output = MSH.run(command)
         if not assert_false(suc, raise_assertion_error=False):
             raise AssertionError(
@@ -68,3 +68,29 @@ class MSH:
     @staticmethod
     def cleanup():
         MSH.HOME.delete()
+
+    @staticmethod
+    def get_all_commands():
+        return [
+            MSHCommand("copy", "cp"),
+            MSHCommand("details", "d"),
+            MSHCommand("edit", "e"),
+            MSHCommand("help", "h"),
+            MSHCommand("new", "n"),
+            MSHCommand("list", "ls"),
+            MSHCommand("manual", "m"),
+            MSHCommand("migrate-scripts"),
+            MSHCommand("move", "mv"),
+            MSHCommand("new", "n"),
+            MSHCommand("print", "p"),
+            MSHCommand("remove", "rm"),
+            MSHCommand("restore"),
+            MSHCommand("sync", "s"),
+            MSHCommand("version", "v"),
+            # ~~add-command-string~~
+        ]
+
+class MSHCommand:
+    def __init__(self, name: str, short_form: str = None):
+        self.name = name
+        self.short_form = short_form